+++ src/pl/plajer/villagedefense3/creatures/v1_12_R1/TankerZombie.java
@@ -15,9 +15,13 @@ public class TankerZombie extends EntityZombie {
     public int damage;
     private float bw;
 
+    public TankerZombie(org.bukkit.World world){
+        this(((CraftWorld) world).getHandle());
+    }
+
     @SuppressWarnings("rawtypes")
-    public TankerZombie(org.bukkit.World world) {
-        super(((CraftWorld) world).getHandle());
+    public TankerZombie(World world) {
+        super(world);
         this.bw = Main.ZOMBIE_SPEED; //Change this to your liking. this is were you set the speed
         this.damage = 15; // set the damage
         //There's also a ton of options of you do this. play around with it