+++ src/pl/plajer/villagedefense3/creatures/v1_8_R3/GolemBuster.java
@@ -20,12 +20,17 @@ import java.util.List;
  * Created by Tom on 14/08/2014.
  */
 public class GolemBuster extends EntityZombie {
+
     public int damage;
     private float bw;
 
+    public GolemBuster(org.bukkit.World world) {
+        this(((CraftWorld) world).getHandle());
+    }
+
     @SuppressWarnings({"rawtypes", "unchecked"})
-    public GolemBuster(World world) {
-        super(((CraftWorld) world).getHandle());
+    public GolemBuster(net.minecraft.server.v1_8_R3.World world) {
+        super(world);
         this.bw = Main.ZOMBIE_SPEED; //Change this to your liking. this is were you set the speed
         this.damage = 15; // set the damage
         //There's also a ton of options of you do this. play around with it