+++ src/pl/plajer/villagedefense3/creatures/v1_8_R3/RidableIronGolem.java
@@ -9,13 +9,13 @@ import java.util.List;
 /**
  * Created by Tom on 17/08/2014.
  */
-public class WorkingWolf extends EntityWolf {
+public class RidableIronGolem extends EntityIronGolem {
 
-    public WorkingWolf(org.bukkit.World world){
+    public RidableIronGolem(org.bukkit.World world) {
         this(((CraftWorld) world).getHandle());
     }
 
-    public WorkingWolf(World world) {
+    public RidableIronGolem(World world) {
         super(world);
 
         List goalB = (List) CreatureUtils.getPrivateField("b", PathfinderGoalSelector.class, goalSelector);
@@ -28,86 +28,88 @@ public class WorkingWolf extends EntityWolf {
         targetC.clear();
 
         this.a(1.4F, 2.9F);
-        ((Navigation) getNavigation()).a(true);
-        this.goalSelector.a(3, new PathfinderGoalLeapAtTarget(this, 0.4F));
-        this.goalSelector.a(4, new PathfinderGoalMeleeAttack(this, 1.0D, true));
-        this.goalSelector.a(5, new PathfinderGoalFollowOwner(this, 1.0D, 10.0F, 2.0F));
-        this.goalSelector.a(2, new PathfinderGoalMeleeAttack(this, EntityZombie.class, 1.5F, false));
+        ((Navigation) getNavigation()).b(true);
+        this.goalSelector.a(1, new PathfinderGoalMeleeAttack(this, 1.0D, true));
+        this.goalSelector.a(2, new PathfinderGoalMoveTowardsTarget(this, 0.9D, 32.0F));
+        this.goalSelector.a(3, new PathfinderGoalMoveThroughVillage(this, 0.6D, true));
         this.goalSelector.a(4, new PathfinderGoalMoveTowardsRestriction(this, 1.0D));
+        this.goalSelector.a(5, new PathfinderGoalOfferFlower(this));
         this.goalSelector.a(6, new PathfinderGoalRandomStroll(this, 0.6D));
         this.goalSelector.a(7, new PathfinderGoalLookAtPlayer(this, EntityHuman.class, 6.0F));
         this.goalSelector.a(8, new PathfinderGoalRandomLookaround(this));
-        this.targetSelector.a(2, new PathfinderGoalNearestAttackableTarget(this, EntityZombie.class, true));
-        this.targetSelector.a(1, new PathfinderGoalHurtByTarget(this, true));
+        this.targetSelector.a(1, new PathfinderGoalDefendVillage(this));
+        this.targetSelector.a(2, new PathfinderGoalHurtByTarget(this, false));
+        this.targetSelector.a(3, new PathfinderGoalNearestAttackableTarget(this, EntityInsentient.class, 0, false, true, IMonster.e));
         this.setHealth(500);
     }
 
-  /*  @Override
-    public void e(float f, float f1) {
+    @Override
+    protected void initAttributes() {
+        super.initAttributes();
+        this.getAttributeInstance(GenericAttributes.FOLLOW_RANGE).setValue(150D);
+    }
 
+    @Override
+    protected void dropDeathLoot(boolean flag, int i) {}
 
-        if (this.passenger != null && this.passenger instanceof EntityLiving) {
+    @Override
+    public void g(float f, float f1) {
+        if(this.passenger != null && this.passenger instanceof EntityLiving) {
             this.lastYaw = this.yaw = this.passenger.yaw;
             this.pitch = this.passenger.pitch * 0.5F;
-            this.b(this.yaw, this.pitch);
-            this.aO = this.aM = this.yaw;
-            f = ((EntityLiving) this.passenger).bd * 0.5F;
-            f1 = ((EntityLiving) this.passenger).be;
-            if (f1 <= 0.0F) {
+            setYawPitch(this.yaw, this.pitch);
+            this.aI = this.aG = this.yaw;
+            f = ((EntityLiving) this.passenger).aZ * 0.5F;
+            f1 = ((EntityLiving) this.passenger).ba;
+            if(f1 <= 0.0F) {
                 f1 *= 0.25F;
             }
 
 
-            if (!this.world.isStatic) {
-                this.i((float) this.getAttributeInstance(GenericAttributes.d).getValue());
-                this.Y = 1.0F;
-                if (f1 <= 0.0F) {
+            if(!this.world.isClientSide) {
+                this.k((float) this.getAttributeInstance(GenericAttributes.MOVEMENT_SPEED).getValue());
+                this.S = 1.0F;
+                if(f1 <= 0.0F) {
                     f1 *= 0.25F;    // Make backwards slower
                 }
                 f *= 0.75F;    // Also make sideways slower
 
                 float speed = 0.12F;    // 0.2 is the default entity speed. I made it slightly faster so that riding is better than walking
-                this.i(speed);    // Apply the speed
-                super.e(f, f1);
+                this.k(speed);    // Apply the speed
+                super.g(f, f1);
             }
 
-            if (this.onGround) {
+            if(this.onGround) {
 
-                this.h(false);
+                this.j(false);
             }
 
-            this.aE = this.aF;
+            this.ay = this.az;
             double d0 = this.locX - this.lastX;
             double d1 = this.locZ - this.lastZ;
             float f4 = MathHelper.sqrt(d0 * d0 + d1 * d1) * 4.0F;
 
-            if (f4 > 1.0F) {
+            if(f4 > 1.0F) {
                 f4 = 1.0F;
             }
 
-            this.aF += (f4 - this.aF) * 0.4F;
-            this.aG += this.aF;
+            this.az += (f4 - this.az) * 0.4F;
+            this.aA += this.az;
         } else {
-            this.W = 0.5F;
-            this.aQ = 0.02F;
-            this.Y = 1.0F;
+            this.S = 0.5F;
+            this.aK = 0.02F;
+            // this.S = 1.0F;
 
-            if (f1 <= 0.0F) {
+            if(f1 <= 0.0F) {
                 f1 *= 0.25F;    // Make backwards slower
             }
             f *= 0.75F;    // Also make sideways slower
 
             float speed = 0.12F;    // 0.2 is the default entity speed. I made it slightly faster so that riding is better than walking
-            this.i(speed);    // Apply the speed
-            super.e(f, f1);
+            this.k(speed);    // Apply the speed
+            super.g(f, f1);
+            this.S = 1.0F;
         }
-    } */
-
-    @Override
-    protected void initAttributes() {
-        super.initAttributes();
-        this.getAttributeInstance(GenericAttributes.FOLLOW_RANGE).setValue(130D);
     }
-
 
 }