+++ src/pl/plajer/villagedefense3/creatures/v1_9_R1/FastZombie.java
@@ -10,12 +10,17 @@ import java.util.Set;
  * Created by Tom on 14/08/2014.
  */
 public class FastZombie extends EntityZombie {
+
     public int damage;
     private float bw;
 
+    public FastZombie(org.bukkit.World world){
+        this(((CraftWorld) world).getHandle());
+    }
+
     @SuppressWarnings("rawtypes")
-    public FastZombie(org.bukkit.World world) {
-        super(((CraftWorld) world).getHandle());
+    public FastZombie(World world) {
+        super(world);
         this.bw = 1.5F; //Change this to your liking. this is were you set the speed
         this.damage = 15; // set the damage
         //There's also a ton of options of you do this. play around with it