+++ src/pl/plajer/villagedefense3/creatures/v1_9_R1/WorkingWolf.java
@@ -11,9 +11,12 @@ import java.util.Set;
  */
 public class WorkingWolf extends EntityWolf {
 
-    public WorkingWolf(org.bukkit.World world) {
-        super(((CraftWorld) world).getHandle());
+    public WorkingWolf(org.bukkit.World world){
+        this(((CraftWorld) world).getHandle());
+    }
 
+    public WorkingWolf(World world) {
+        super(world);
 
         Set goalB = (Set) CreatureUtils.getPrivateField("b", PathfinderGoalSelector.class, goalSelector);
         goalB.clear();